Core/BattlePets Fix Linux compilation
